Fix HC bugs
This commit fixes two bugs in HoloClean:
1) HC cannot clean datasets that contain a column named "index"
(without the "). To fix this, the dataframe-method to_records()
needs to be passed the index=False attribute.
2) HC fails to properly parse unary DCs. This is due to a bug in
load_denial_constraints(), in the DenialConstraint class:
contains_operation() returns either None if no operation is contained
in a split, or the position of the operation in the split. When
if (contains_operation(split)) is evaluated though and the operation
is in position 0 of the split, the expression equals False, and
the operation is incorrectly identified as a tuple_name. This is fixed now.