Dockerize, small fix for llm value correction
I ended up putting dependencies into a requirements.txt file, which
seems like a contradiction to the conda-approach. But I did not figure
out a way to work with the arm-environment.yml file on a linux-machine.
Oh well, I guess the code works.

This also fixes a small bug in the llm value-correction model to ignore
NULLs better.